statbot_backend/requirements.txt
# statbot_backend/requirements.txt
langchain
langchain-experimental
langchain-groq
pandas
matplotlib
python-dotenv
tabulate
fastapi
uvicorn
websockets
python-multipart
